requirements.txt
Django==1.11.4
djangorestframework==3.6.3
jsonfield==2.0.2
requests==2.18.3
channels==1.1.8
asgi_redis==1.4.3
pandas==0.21.0
django-prettyjson==0.3.0
ipython==5.5.0
kafka-python==1.4.2
psycopg2==2.7.5